update stories, generate autodocs
